ibacm: add support for path lookup by LID/GID
Support path record lookup based on DLID or DGID.  This allows
ibacm to be used purely for path record caching, with address
resolution being done by a user.  This is needed to support
clients that exchange LID/GID information out of band before
connecting, or that make use of ipoib for address resolution.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>