infiniband-diags: Make sure buffers are set to zero before calling xxx_query_via
where xxx is smp and pma. This is a cosmetic change based on
IBA 1.2.1 vol 1 p.725 line 9:

When constructing a response MAD that contains all or part of
the corresponding request MAD, it is acceptable to include the
contents of reserved fields in the request MAD in the response
MAD without regard to their content. That is, such fields need not
be set to zero in the response MAD.

Signed-off-by: Aviad Yehezkel <aviadye@mellanox.co.il>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>