rsocket: Succeed SO_LINGER socket option
Succeed calls to set the SO_LINGER socket option.  We don't
actually implement SO_LINGER semantics because we never place
an rsocket into a timewait state.  Unlike socket behavior,
we do wait for all pending data to be transferred by the hardware.
This is done so that the disconnect message can be sent over
the rsocket connection.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>