rdma_xserver/client: Fix man page formatting
Putting 'r' at the beginning of a line in the nroff source for man pages
is confusing to nroff because lines that start with a single quote
character ' or a dot character . are treated as control lines, which is
not what's intended here.  Some of the man page text ends up left out of
the formatted output.

Fix this by just wrapping the text slightly differently in the source
(which doesn't matter since nroff reflows the text anyway).  Also add a
missing ".TP" so that the -p and -c options are not run together in the
formatted output.

Signed-off-by: Roland Dreier <roland@purestorage.com>
Signed-off-by: Sean Hefty <sean.hefty@intel.com>