flush bug can cause application crash
This fixes a problem in the flush logic that could occur if the
current position is near the end of the queue and the 'count' being
retired plus this position was greater than the queue size. In this
case the index wasn't being wrapped and the associated swcqe pointer
would point past the end of the CQ.

Asserts were added to check other possible wrap-error paths.
    
This bug was causing mbw to occasionally crash when flushing the qp.

Signed-off-by: Steve Wise <swise@opengridcomputing.com>