fixed bug where sq.udb was getting overwritten.
There was an erroneous statement in create_rc_qp() which caused the
sq.udb va to be overwritten with the rq.udb va only for on-chip queues.
This error only affected unmapping of that mapped va when destroying
a qp.  In the presence of multiple threads creating/destroying qps, the
erroneously unmapped va could have been assigned to another qp mapping.
This causes intermittent seg faults as libcxgb4 tries to access memory
that has been unmapped in error.

Signed-off-by: Steve Wise <swise@opengridcomputing.com>