Infrastructure to support verbs extensions
Add infrastructure to support extended verbs capabilities in a
forward/backward manner.

Support for extensions is determined by the provider calling
verbs_register_driver in place of ibv_register_driver.  When
extensions are enabled, ibverbs sets the current alloc_context /
free_context device operations to NULL.  These are used to indicate
that the struct ibv_device may be cast to struct verbs_device.

With extensions, ibverbs allocates the ibv_context structure and calls
into the provider to initialize it.  The init call is part of the
verbs_device struct.

The abi_compat field of struct ibv_context is used to determine
support of verbs extensions.  As a result, support for ABI version < 2
is removed (corresponds to kernel releases 2.6.11-2.6.14 no longer
being supported).  The lowest ABI now supported is 3 (really 4 since
2.6.15 was ABI 4, I don't see that ABI 3 was in a release).

Acked-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Tzahi Oved <tzahio@mellanox.com>
Signed-off-by: Sean Hefty <sean.hefty@intel.com>
Signed-off-by: Roland Dreier <roland@purestorage.com>