Introduce XRC domains (XRCDs)
XRC introduces several new concepts and structures, one of which is
the XRC domain.

XRC domains: XRCDs are a type of protection domain used to associate
shared receive queues with XRC queue pairs.  Since XRCDs are meant to be
shared among multiple processes, we introduce new APIs to open/close
XRCDs.

The user-kernel ABI is extended to account for opening/closing XRCDs.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Roland Dreier <roland@purestorage.com>