Add support for XRC SRQs
XRC support requires the use of a new type of SRQ: XRC shared receive
queues.  XRC SRQs are similar to normal SRQs, except that they are bound
to an XRCD, rather than to a protection domain.  Based on the current
spec and implementation, they are only usable with XRC QPs. To support
XRC SRQs, we define a new srq_init_attr structure to include an SRQ type
and other needed information.

The user-kernel ABI is also updated to allow creating extended SRQs.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Roland Dreier <roland@purestorage.com>