Add support for XRC QPs
XRC queue pairs: XRC defines two new types of QPs.  The initiator, or
send-side, XRC QP behaves similar to a send-only RC QP.  XRC send QPs
are managed through the existing QP functions.  The send_wr structure
is extended in a backwards compatible way to support posting sends on
a send XRC QP, which require specifying the remote XRC SRQ.

The target, or receive-side, XRC QP behaves differently than other
implemented QPs.  A recv XRC QP can be created, modified, and
destroyed like other QPs through the existing calls.  The
qp_init_attr structure is extended for XRC QPs.

Because XRC recv QPs are bound to an XRCD, rather than a PD, they are
intended to be used among multiple processes.  Any process with access
to an XRCD may allocate and connect an XRC recv QP.  The actual XRC
recv QP is allocated and managed by the kernel.  If the owning process
explicitly destroys the XRC recv QP, it is destroyed.  However, if the
XRC recv QP is left open when the user process exits or closes its
device, then the lifetime of the XRC recv QP is tied to the lifetime
of the XRCD.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Roland Dreier <roland@purestorage.com>