truescale.cmds: initialize serdes for mezzanine cards
A number of QLogic mezzanine cards need to know which slot they are
connected to in order to set the serdes parameters correctly.
This is because there is no QSFP EEPROM to indicate the trace length
to the switch. The fix is to use dmidecode to get the blade slot number
and set the serdes accordingly.

Signed-off-by: Ralph Campbell <ralph.campbell@qlogic.com>