iw_cxgb4/libcxgb4: swsq "signaled" state needs to mind sq_sig_all.
When a qp is created, the ULP can request that all sq wrs be signaled.
The swsqe entries need to take this into account when marking the entry
as signaled.

This bug will cause read request CQEs to be silently discarded, and can
cause flush problems because the swsqe entry will be mismarked for qps
created with sq_sig_all set in the qp init attributes.

Signed-off-by: Steve Wise <swise@opengridcomputing.com>