ibacm: Convert acm_addr_index call to return address
Rename acm_addr_index to acm_addr_lookup and have it
return the acm_address that matches the source address,
rather than the index that the address is located.
The index is an internal limitation of the acm core
and the acm_address is usable by the providers.

Change acm_addr_lookup to take the endpoint as input
in place of the provider ep.  acm_addr_lookup is
intended to be exported by the acm core.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>