ibacm: Define provider endpoint identifier
Introduce a header file that will be included by providers that
will define the interface between the ibacm core layer and the
providers.

Define an endpoint identifier that is associated with a set of
addresses.  Providers will be given this structure in requests.
The endpoint is defined as a <device, port, pkey> tuple, though
it may be used with "any" values in order to identify any
port on a device and/or any pkey on a port.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>