ibacm: Provide correct addr_len for IPv6
When reading in addresses from the address file, the
address string is converted into either an IPv4, IPv6,
or hostname 'address'.  In the case of IPv6, the length
of the address is set to the maximum supported value.
This can result in extra bytes being set beyond the
actual IPv6 address, which are then copied into the
endpoint address.  Fix this by setting the size
of the IPv6 correctly.  When the address is copied into
the endpoint address, the extra bytes are then ignored.

Problem reported by: Kaike Wan <kaike.wan@intel.com>

Signed-off-by: Sean Hefty <sean.hefty@intel.com>