Add receive flow steering support
The RDMA stack allows for applications to create IB_QPT_RAW_PACKET
QPs, which receive plain Ethernet packets -- specifically packets that
don't carry any QPN to be matched by the receiving side.  Applications
using these QPs must be provided with a method to program some
steering rule into the HW so packets arriving at the local port can be
routed to them.

In a similar manner, when the device supports flow streeing, IB UD QPs
created by IPoIB allow userspace applications to steer specific TCP/IP
flows to their QPs.

This patch adds ibv_create_flow(), which allow providing a flow
specification for a QP.  When there's a match between the
specification and a received packet, the packet is forwarded to that
QP, in a the same way one uses ibv_attach_mcast() for IB UD multicast
handling.

Flow specifications are provided as instances of struct
ibv_flow_spec_yyy, which describes L2, L3 and L4 headers.  Currently
specs for Ethernet, IPv4, TCP and UDP are defined.  Flow specs are
made of values and masks.

The input to ib_create_flow() is a struct ib_flow_attr, which contains
a few mandatory control elements and optional flow specs.

 struct ibv_flow_attr {
	uint32_t comp_mask;
	enum ibv_flow_attr_type type;
	uint16_t size;
	uint16_t priority;
	uint8_t  num_of_specs;
	uint8_t  port;
	uint32_t flags;
	/* Following are the optional layers according to user request
	 * struct ibv_flow_spec_xxx [L2]
	 * struct ibv_flow_spec_yyy [L3/L4]
	 */
 };

These flow specs are defined and used in a way which allows adding new
spec types without kernel/user ABI change, just with a little API
enhancement which defines the newly added spec.

The flow spec structures are defined with TLV (Type-Length-Value)
entries, which allows calling ib_create_flow() with a list of variable
length of optional specs.

For the actual processing of ibv_flow_attr the kernel uses the number
of specs and the size mandatory fields along with the TLV nature of
the specs.

The returned value from ibv_create_flow() is a struct ibv_flow, which
contains a handle provided by the kernel to be used when calling
ibv_destroy_flow().

The ib_flow_attr enum type supports usage of flow steering for
promiscuous and sniffer purposes:

    IBV_FLOW_ATTR_NORMAL - "regular" rule, steering according to rule
        specification

    IBV_FLOW_ATTR_ALL_DEFAULT - default unicast and multicast rule,
        receive all Ethernet traffic which isn't steered to any QP

    IBV_FLOW_ATTR_MC_DEFAULT - same as IB_FLOW_ATTR_ALL_DEFAULT but
        only for multicast

ALL_DEFAULT and MC_DEFAULT rules options are valid only for Ethernet
link type.

Signed-off-by: Hadar Hen Zion <hadarh@mellanox.com>
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: Matan Barak <matanb@mellanox.com>
Signed-off-by: Roland Dreier <roland@purestorage.com>