ibacm: Open a provider ep only if an address is assigned to the core ep.
The function acm_assign_ep_names() should not return 0 when there is no valid
address assigned to the core ep. In this case, the provider's
ep should not be up (or acmp_ep_up() should not be called). However, when
there is at least one address assigned, 0 should be returned so that the
core endpoint and the provider endpoint should be created.

Signed-off-by: Kaike Wan <kaike.wan@intel.com>
Signed-off-by: Sean Hefty <sean.hefty@intel.com>