Rename truescale.cmds to truescale-serdes.cmds and remove outdated functionality
This patch renames the truescale.cmds file to truecale-serdes.cmds. It no longer
needs to do the IRQ affinity stuff as this was for older kernels which did not
do irqbalance correctly.

This patch also removes the mounting of /ipathfs. This is not strictly required
and can be mounted on demand if the user needs to access.

Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>