ibacm: Define config, bin, and prov lib path in configure.ac file.
The reason behind this change is to make the ibacm
module works properly under normal build (./autogen.sh, ./configure,
make, make
install) and rpm build (./autogen.sh, ./configure, make dist, rpmbuild).
Without this change, the rpmbuild will not work properly:
 (1) For normal build, the files will be installed under:
          /usr/local/sbin: ibacm
          /usr/local/bin: ib_acme
          /usr/local/etc/rdma/: configuration files.

 (2) For rpmbuild, the files will be installed under:

          /usr/sbin: ibacm
          /usr/bin: ib_acme
          /etc/rdma/: configuration files.

The configure.ac will be used in both cases to set up the path correctly.
This is necessary when we need to load the provider library from the
appropriate path.


Signed-off-by: Kaike Wan <kaike.wan@intel.com>
Signed-off-by: Sean Hefty <sean.hefty@intel.com>