ibacm: Centralize provider SA access in core.
This patch is the first of a series to centralize the SA access for all
providers.

Here is the description of the approach:
(1) A provider will call the core to allocate a request and initialize
    the request (including the callback function);
(2) Provider calls the core to send the SA request;
(3) The core SA handler thread will poll for any SA response;
(4) When a SA response is received, the core SA handler will find the
    matching request and call the corresponding provider callback
    function;
(5) The provider callback will handle the SA response and free the
    SA request.

The approach has the following advantages:
(1) No code duplication in the provider for SA access;
(2) No need for a separate thread in the provider to wait for SA response.

Signed-off-by: Kaike Wan <kaike.wan@intel.com>
Signed-off-by: Sean Hefty <sean.hefty@intel.com>