hfiverbs: add autogen.sh files to fix cvs import
The current dist tar ball cannot be built unless
the ibverbs rpm is installed.

This patch adds the side effects of autogen.sh
to the git repo.

It also builds the dist tarball outside of the Makefile
and converts to current versioning convention for the
driver and psm.

INTERNAL: PR 124251 - Support allow CVS import without having verbs installed

Reviewed-by: John Fleck <john.fleck@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>