riostream: Only verify last data transfer
Data verification will fail when running the bandwidth
tests or the transfer count is > 1.  The issue is that
subsequent writes by the initiator side will overwrite
the data in the target buffer before the receiver can
verify that it is correct.

To fix this, only verify that the data in the buffer
is correct after the last transfer has completed.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>