libcxb4: use wc_wmb() instead of wmb().
The user mode RDMA macro for wmb() is not sufficient to provide the
fencing needed.  This was causing corrupted packets when running multiple
WD-UDP sockperf processes.  The correct "sfence" instruction is provided
with the wc_wmb() macro.  See /usr/include/infiniband/arch.h.

Note:  The kernel implementation of wmb() is sufficient.  This is a
libcxgb4 issue only...

Signed-off-by: Steve Wise <swise@opengridcomputing.com>